Remove Hashing#toHexString:
This method does not belong in `Hashing`,
and used in a single place only.
