+++ CHANGELOG.md
@@ -5,7 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+<!-- Use the following sections from the spec: http://keepachangelog.com/en/1.0.0/
+  - Added for new features.
+  - Changed for changes in existing functionality.
+  - Deprecated for soon-to-be removed features.
+  - Removed for now removed features.
+  - Fixed for any bug fixes.
+  - Security in case of vulnerabilities. -->
+
 ## [Unreleased]
+
+### Removed
+- `Hashing#toHexString`. (#379)
 
 ## 0.2 - 2018-07-23
